back/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
envparse==0.2.0
numpy==1.18.4
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
